requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-allauth==0.57.0
django-auditlog==2.3.0
django-crispy-forms==2.0
django-embed-video==1.4.9
django-environ==0.11.2
django-sass-processor==1.2.2
idna==3.4
libsass==0.22.0
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.5
